Re-organize project structure, and add docker config files for dockerized building of the firmware.
