--- prawn_do.pio
@@ -1,112 +1,0 @@
-.program prawn_do
-
-.define public TRIGGER_PIN 16 ; hard code trigger to pin 16
-.define public OUTPUT_PIN_BASE 0 ; first pin to output from
-.define public OUTPUT_WIDTH 16 ; number of pins to output from
-
-; Immediately output the first entry so the outputs could be manually triggered
-;out pins, 32
-; pop off dummy reps value from OSR
-;out X, 32
-
-start:
-	;wait 1 gpio TRIGGER_PIN
-	out X, 32 ; check first value for software or hardware start
-	jmp !X new_output ; if 0, software start sequence
-
-	wait 1 gpio TRIGGER_PIN ; else hardware start sequence
-; Main Execution Loop:
-.wrap_target
-new_output:
-	; Latest fifo entry is autopulled into the OSR (check C code below)
-	out pins, 32 ; Bit-banging from the data stored in the OSR to the pins
-
-	; Latest fifo entry is autopulled into the OSR
-	out X, 32 ; store the number of repetitions in the X scratch register 
-
-	jmp !X wait_end ; If the number of repetitions inputted is zero or if the OSR
-			        ; is empty, this will jump out to then determine if there is
-			        ; an indefinite wait or a full stop
-executing_pulse:
-	jmp X-- executing_pulse ; Decrementing the X value and looping back
-	; Automatically loop back to the wrap_target to continue execution
-.wrap
-; The program reaches this point if the number of repetitions is equal to zero 
-; or if the OSR is empty in the main execution loop
-wait_end:
-	out Y, 32 ; Store the next OSR entry into the Y scratch register
-
-	out X, 32 ; Store the number of reps in the X scratch register
-
-	jmp !X end ; If reps equals zero, then jump to end, otherwise indefinite wait
-; This section of the code is used for waits that can only be exited by a 
-; hardware trigger
-indefinite_wait:
-	wait 1 gpio TRIGGER_PIN    ; Wait for a hardware trigger
-
-	mov pins, Y ; Move the output word stored in Y to the pins
-
-	nop ; Skip one cycle for alignment purposes
-
-; Loop for each rep until the rep counter is zero
-end_indefinite_wait_loop:
-	jmp X-- end_indefinite_wait_loop
-
-	; Jump to the top and continue executing commands
-	jmp new_output
-
-; send isr=0 data to rx fifo to signal program end
-end:
-	irq wait 0 rel
-; The purpose of this section is to continuously loop to prevent wrapping
-end_loop:
-	jmp end_loop ; Continuously loop
-
-
-				   
-% c-sdk {
-pio_sm_config prawn_do_program_init(PIO pio, uint state_machine, float div, 
-	uint offset){
-	// Set pin direction of output pins to outputs
-	pio_sm_set_consecutive_pindirs(pio, state_machine,
-								   prawn_do_OUTPUT_PIN_BASE,
-								   prawn_do_OUTPUT_WIDTH, true);
-
-	// Set pin direction of trigger pin to input
-	pio_sm_set_consecutive_pindirs(pio, state_machine, prawn_do_TRIGGER_PIN, 1, false);
-
-	// Initialize gpio for output pins
-	for(uint i = 0; i < prawn_do_OUTPUT_WIDTH; i++){
-		pio_gpio_init(pio, prawn_do_OUTPUT_PIN_BASE + i);
-	}
-
-	// Initialize gpio for trigger pin
-	pio_gpio_init(pio, prawn_do_TRIGGER_PIN);
-
-	// Get config for pio state machine
-	pio_sm_config config = prawn_do_program_get_default_config(offset);
-
-	// Set output pins of config to output pins
-	sm_config_set_out_pins(&config, prawn_do_OUTPUT_PIN_BASE, prawn_do_OUTPUT_WIDTH);
-	// Set input pin of config to trigger pin
-	sm_config_set_in_pins(&config, prawn_do_TRIGGER_PIN);
-
-	// Setup automatic shift on output.
-	// When 32 bits are outputted anywhere within the PIO code,
-	// another entry is pulled from the fifo (the fifo is then refilled via dma)
-	sm_config_set_out_shift(&config, true, true, 32);
-
-	// Clear IRQ flag before starting, and make sure flag doesn't actually
-    // assert a system-level interrupt (we're using it as a status flag)
-    pio_set_irq0_source_enabled(pio,
-		                        (enum pio_interrupt_source) ((uint) pis_interrupt0 + state_machine),
-								false);
-	pio_set_irq1_source_enabled(pio,
-		                        (enum pio_interrupt_source) ((uint) pis_interrupt0 + state_machine),
-								false);
-    pio_interrupt_clear(pio, state_machine);
-
-	// Complete initialization of state machine.
-	pio_sm_init(pio, state_machine, offset, &config);
-}
-%}
